requirements.txt
Flask==3.0.3
pypdf==4.2.0
Werkzeug==3.0.3
SQLAlchemy==2.0.30
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
flask-pytest==0.0.5
reportlab==4.2.0
load-dotenv==0.1.0
alembic==1.13.1
factory-boy==3.3.0
boto3==1.34.125
botocore==1.34.125
moto[ec2,s3,all]==5.0.9
schedule==1.2.2